Реализована функционально CategoryService, половина EventService

+++ main-service/src/main/java/ru/practicum/ewm/category/service/CategoryServiceImpl.java
@@ -2,16 +2,21 @@ package ru.practicum.ewm.category.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewm.category.Category;
 import ru.practicum.ewm.category.CategoryRepository;
 import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.category.dto.CategoryMapper;
 import ru.practicum.ewm.category.dto.NewCategoryDto;
+import ru.practicum.ewm.event.EventRepository;
+import ru.practicum.ewm.exceptions.ConflictException;
 import ru.practicum.ewm.exceptions.NotFoundException;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -19,33 +24,79 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
-        return null;
+        log.info("Проверка наличия категории с name {}", newCategoryDto.getName());
+        checkUniqueName(newCategoryDto.getName());
+        log.info("Добавление новой категории {}", newCategoryDto);
+        Category category = categoryRepository.save(categoryMapper.toCategory(newCategoryDto));
+        log.info("Категория успешно добавлена с id {}", category.getId());
+        return categoryMapper.toCategoryDto(category);
     }
 
     @Override
     public void deleteCategory(int catId) {
         log.info("Удаление категории с id {}", catId);
-        Category category = categoryRepository.findById(catId).orElseThrow(
-                () -> new NotFoundException("Категория с id = " + catId + " не найден"));
+        Category category = checkCategoryExist(catId);
+
+        log.info("Проверка наличия событий привязанных к катеории");
+        if(eventRepository.findByCategoryId(catId)) {
+            throw new ConflictException("Нельзя удалить категорию, с которой связаны события");
+        }
+
         categoryRepository.deleteById(catId);
         log.info("Категория успешно удалена");
     }
 
     @Override
     public CategoryDto updateCategory(int catId, CategoryDto categoryDto) {
-        return null;
+        log.info("Проверка наличия категории перед обновлением");
+        Category category = checkCategoryExist(catId);
+
+        if (categoryDto.getName() != null && !categoryDto.getName().equals(category.getName())) {
+            log.info("Проверка наличия категории с name {}", categoryDto.getName());
+            checkUniqueName(categoryDto.getName());
+        }
+
+        log.info("Обновление категории");
+        categoryDto.setId(category.getId());
+        Category savedCategory = categoryRepository.save(categoryMapper.toCategory(categoryDto));
+
+        log.info("Категория успешно добавлена с id {}", savedCategory.getId());
+        return categoryMapper.toCategoryDto(savedCategory);
     }
 
     @Override
     public Collection<CategoryDto> getAllCategories(int from, int size) {
-        return List.of();
+        log.info("Получение списка категорий с from {}, size {}", from, size);
+        PageRequest page = PageRequest.of(from, size, Sort.by("category_id").ascending());
+        List<Category> categories = categoryRepository.findAll(page).getContent();
+
+        log.info("Категории успешно найдены");
+        return categories.stream()
+                .map(CategoryMapper::toCategoryDto)
+                .collect(Collectors.toList());
     }
 
     @Override
     public CategoryDto getCategoryById(int catId) {
-        return null;
+        log.info("Поиск категории с id {}", catId);
+        Category category = checkCategoryExist(catId);
+        log.info("Категория с id {} успешно найдена {}", catId, category);
+        return categoryMapper.toCategoryDto(category);
+    }
+
+    private Category checkCategoryExist(int catId) {
+        Category category = categoryRepository.findById(catId).orElseThrow(
+                () -> new NotFoundException("Категория с id = " + catId + " не найдена"));
+        return category;
+    }
+
+    private void checkUniqueName(String name) {
+        if (categoryRepository.findByName(name)) {
+            throw new ConflictException("Категория с name: " + name + " уже существует. Добавление невозможно");
+        }
     }
 }
